pyproject.toml
[project]
name = "rcf-consultant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "fastapi[standard]>=0.124.0",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "sqlmodel>=0.0.27",
    "uvicorn[standard]>=0.38.0",
]
